support setting lightingTheme via MQTT
also publish the proper metadata for it to work via Home Assistant.
note that this requires publishing an additional state topic that
includes both the on/off state and the theme in a single topic
(HASS can't deal with a single topic for commands and separated
topics for state - it either needs to be a single topic for all
attributes on both state and command, or it needs to be separate
topics for each attribute on both state and command).