fix MQTT LWT leaving status as "lost"
because we weren't recording that the LWT was "sent" on disconnect,
the "don't send if it hasn't changed" logic was never re-sending status.
so on an error, go through and simulate a send on any LWT topics

also, it will reconnect right away, but the next "controller" event
(the status message) won't be sent until the next minute; force
one to be sent right away so that other MQTT clients can know it
was reconnected.